# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Beyblade API",
    "version": "1.0.0"
  },
  "paths": {
    "/beyblades": {
      "get": {
        "summary": "Retrieve a list of Beyblades",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "The maximum number of Beyblades to retrieve",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "The number of Beyblades to skip",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Beyblade"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/beyblades/{id}": {
      "get": {
        "summary": "Retrieve a specific Beyblade by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The ID of the Beyblade to retrieve",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BeybladeById"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Beyblade": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "reference": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "url": {
            "type": "string"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "err": {
            "type": "string"
          }
        }
      },
      "BeybladeById": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Beyblade"
          },
          {
            "type": "object",
            "properties": {
              "fandomUrl": {
                "type": "string"
              },
              "imageUrl": {
                "type": "string"
              },
              "productInfo": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "series": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "system": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "spinDirection": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              },
              "parts": {
                "type": "object",
                "additionalProperties": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        ]
      }
    }
  }
}